backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "yapflows"
version = "2.0.0"
description = "Lightweight self-hosted personal AI assistant framework"
readme = "../README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "websockets>=12.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.26.0",
    "aiofiles>=23.0.0",
    "apscheduler>=3.10.0",
    "PyYAML>=6.0",
    "playwright>=1.40.0",
    "playwright-stealth>=2.0.0",
    "websockify>=0.13.0",
    "strands-agents>=0.1.0",
    "strands-agents-tools>=0.1.0",
    "litellm>=1.0.0",
    "python-frontmatter>=1.1.0",
    "python-telegram-bot>=21.0",
    "jinja2>=3.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
markers = [
    "integration: Integration tests with mocked external services (zero token usage)",
    "external: Tests requiring real external services and API keys (costs tokens!)",
    "docker: Tests requiring Docker",
    "slow: Slow tests (e.g., playwright browser install)",
]
addopts = "-v -m 'not external'"
